ingestion/requirements.txt
psycopg2-binary
python-dotenv
numpy
faiss-cpu
scipy
